restructured folder hierarchy to experiment with a feature based folder structure
